requirements.txt
# ============================================================================
# Satellite Property Valuation - Requirements
# ============================================================================

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                     # PyTorch Image Models (ConvNeXt, EfficientNet)

# Computer Vision
albumentations>=1.3.0           # Image augmentations
opencv-python-headless>=4.8.0   # Image processing
Pillow>=10.0.0                  # Image loading

# Data Science
pandas>=2.0.0
numpy>=1.24.0,<2.0.0           # NumPy 1.x for compatibility
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0                    # Progress bars
openpyxl>=3.1.0                 # Excel file reading
python-dotenv>=1.0.0            # Environment variables

# Machine Learning (for Approach 1)
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Experiment Tracking (optional, for Approach 2)
mlflow>=2.8.0

# Development
pytest>=7.4.0                   # Testing
black>=23.7.0                   # Code formatting
flake8>=6.1.0                   # Linting
